Berechnung max HR mittels Birthdate umgesetzt

+++ my_functions.py
@@ -4,9 +4,9 @@ def estimate_max_hr(age_years : int , sex : str) -> int:
   See https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4124545/ for different formulas
   """
   if sex == "male":
-    max_hr_bpm =  223 - 0.9 * age_years
+    max_hr_bpm =  int(223 - 0.9 * age_years)
   elif sex == "female":
-    max_hr_bpm = 226 - 1.0 *  age_years
+    max_hr_bpm = int(226 - 1.0 *  age_years)
   else:
     # der input() öffnet ein Eingabefenster für den Nutzer und speichert die Eingabe
     max_hr_bpm  = input("Enter maximum heart rate:")
